Further misc refactoring of collab editor files
